% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackProjectFunctions.r
\name{abb}
\alias{abb}
\title{Take abbreviation -- translates a string into its abbreviation}
\usage{
abb(q, suffix = "", prefix = "", rm_articles = F)
}
\description{
Take abbreviation -- translates a string into its abbreviation
}
