% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/no_commas.r
\name{data.frame}
\alias{data.frame}
\title{Make it acceptable for the data.frame() function to end in a comma}
\usage{
data.frame(...)
}
\description{
Make it acceptable for the data.frame() function to end in a comma
}
