% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dichNA.R
\name{dichNA}
\alias{dichNA}
\title{Dichotomize with NAs}
\usage{
dichNA(data, variables, add.data = F)
}
\arguments{
\item{data}{Dataset from which you are pulling variables}

\item{variables}{List variable names in a string}
}
\description{
This function dichotomizes variables while including a column for NAs.
}
\examples{
dichNA(data=y,variables=c("age","sex"))
}
\keyword{Dichotomize}
