% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackProjectFunctions.r
\name{fnFinder}
\alias{fnFinder}
\title{Find the first name. Split up all the words by spaces. If there is a comma, it takes the first word after the comma (Last, First). Otherwise, it takes the first name.}
\usage{
fnFinder(x)
}
\description{
Find the first name. Split up all the words by spaces. If there is a comma, it takes the first word after the comma (Last, First). Otherwise, it takes the first name.
}
