% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackProjectFunctions.r
\name{grepl0}
\alias{grepl0}
\title{Advanced grepl -- this was an attempt to make grepl a bit more usable. It doesn't work well.}
\usage{
grepl0(pattern, string)
}
\description{
Advanced grepl -- this was an attempt to make grepl a bit more usable. It doesn't work well.
}
