% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackProjectFunctions.r
\name{hypFinder}
\alias{hypFinder}
\title{Find the hyphen. Look for a hyphenated name. If a hyphenated name exists (e.g., Smith-Jones), it can return the word before the hyphen (if before=T, Smith) or after (if before=F, Jones).}
\usage{
hypFinder(x, before = T)
}
\description{
Find the hyphen. Look for a hyphenated name. If a hyphenated name exists (e.g., Smith-Jones), it can return the word before the hyphen (if before=T, Smith) or after (if before=F, Jones).
}
