% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackProjectFunctions.r
\name{lnFinder}
\alias{lnFinder}
\title{Find the last name. Split up all words by spaces. If there is a comma, it takes the first word (Last, First). Otherwise, it takes the last name.}
\usage{
lnFinder(x)
}
\description{
Find the last name. Split up all words by spaces. If there is a comma, it takes the first word (Last, First). Otherwise, it takes the last name.
}
