% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackProjectFunctions.r
\name{more.than.1.league}
\alias{more.than.1.league}
\title{There is more than one unique value in the vector}
\usage{
more.than.1.league(x)
}
\description{
There is more than one unique value in the vector
}
