% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackProjectFunctions.r
\name{more.than.two.diff.league}
\alias{more.than.two.diff.league}
\title{There are more than two elements different from the most common element in a vector}
\usage{
more.than.two.diff.league(x)
}
\description{
There are more than two elements different from the most common element in a vector
}
