% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackProjectFunctions.r
\name{most.common.league}
\alias{most.common.league}
\title{Most common league -- the most common vector in a list}
\usage{
most.common.league(x)
}
\description{
Most common league -- the most common vector in a list
}
