% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newZelig.R
\name{newZelig}
\alias{newZelig}
\title{NewZelig}
\usage{
newZelig(ind_var_group, ind_var_value = 1, zero_out = NULL, dep_var,
  other_variables, factor_variables, .id = NULL, .weight = NULL,
  .nest = NULL, .strata = NULL, total_variables = c(dep_var,
  zero_out, other_variables, substr(.id, 1, 10)[2], substr(.weight, 1,
  10)[2], substr(.strata, 1, 10)[2]), dat, sex_interaction = F,
  totalData = dat[, total_variables], seed = 1234,
  function_family = binomial(link = "logit"))
}
\arguments{
\item{ind_var_group}{The factor variable for which you want an estimate}

\item{ind_var_value}{The value of ind_var_group that will be evaluated. Default set at 1 (100%).}

\item{zero_out}{The factor variable for which we want to zero-out all other possibilities.}

\item{dep_var}{Dependent variable}

\item{other_variables}{What other confounders do you want to include in the model}

\item{factor_variables}{Which of the variables should be considered categorical?}

\item{sex_interaction}{Include if you want to interact the independent variable with sex. Default set to F.}

\item{data}{Dataset you want to use}
}
\description{
This function provides simulation-based risk ratios from a regression that involves factors.
}
\examples{

}
\keyword{Ratios}
\keyword{Risk}
\keyword{Zelig}
