% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newZeligX.R
\name{newZeligIntX}
\alias{newZeligIntX}
\title{Dichotomize with NAs}
\usage{
newZeligIntX(ind_var_group1, ind_var_value1, ind_var_group2,
  ind_var_value2, zero_out, dep_var, other_variables, factor_variables,
  total_variables = na.omit(c(dep_var, zero_out, other_variables)),
  dat = k, sex_interaction = F, totaldata = y[, total_variables])
}
\arguments{
\item{data}{Dataset from which you are pulling variables}

\item{variables}{List variable names in a string}
}
\description{
This function dichotomizes variables while including a column for NAs.
}
\examples{
dichNA(data=y,variables=c("age","sex"))

}
\keyword{Dichotomize}
