% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cd.r
\name{numbers_only}
\alias{numbers_only}
\title{Change Directory}
\usage{
numbers_only(x)
}
\arguments{
\item{directory}{The directory you want to set.}
}
\description{
This function allows you to set a directory, whether it exists or not.
}
\examples{
cd()
}
\keyword{directory}
