% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackProjectFunctions.r
\name{numWords}
\alias{numWords}
\title{How many words in this string? (Based upon a split of " ")}
\usage{
numWords(str1, split = " ")
}
\description{
How many words in this string? (Based upon a split of " ")
}
