% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackProjectFunctions.r
\name{over.80.percent}
\alias{over.80.percent}
\title{Over 80 percent of the elements in a vector are the same}
\usage{
over.80.percent(x)
}
\description{
Over 80 percent of the elements in a vector are the same
}
