% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cd.r
\name{predict_expand}
\alias{predict_expand}
\title{Predict and Expand}
\usage{
predict_expand(reg, new.time)
}
\description{
Takes the non-intersection of two sets
}
\examples{
predict_expand(reg = lm(outcome ~ time), new.time = 1990:2020)
}
