% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cd.r
\name{read_feather0}
\alias{read_feather0}
\title{Change Directory}
\usage{
read_feather0(data, ...)
}
\arguments{
\item{directory}{The directory you want to set.}
}
\description{
This function allows you to set a directory, whether it exists or not.
}
\examples{
cd()
}
\keyword{directory}
