% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackProjectFunctions.r
\name{remove_articles}
\alias{remove_articles}
\title{Remove articles and other small words (and/of/for/in/the) from text -- helps with abbreviations}
\usage{
remove_articles(q)
}
\description{
Remove articles and other small words (and/of/for/in/the) from text -- helps with abbreviations
}
