% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss.R
\name{ss}
\alias{ss}
\title{Summary Statistics}
\usage{
ss(x, name = "")
}
\arguments{
\item{x}{Data vector}

\item{name}{The name of the vector you're interested in}
}
\description{
This function creates a table of summary statistics.
}
\examples{
with(y,do.call(rbind.data.frame,list(ss(age,"age"),ss(cohort),ss(gender), ss(smoker),ss(sbp,"sbp"),ss(choliu,"choliu"),ss(cvd,"cvd"))))
}
\keyword{statistics}
\keyword{summary}
