% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackProjectFunctions.r
\name{te}
\alias{te}
\title{Text exists: Is this text in this vector? For example, is "A" in LETTERS?}
\usage{
te(data, text)
}
\description{
Text exists: Is this text in this vector? For example, is "A" in LETTERS?
}
