% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testingFunctions.R
\name{testInside}
\alias{testInside}
\title{Test what's in the list}
\usage{
testInside(
  data,
  col = "schoolid",
  string = "http://co.milesplit.com/teams/<TEST>",
  sample = 1,
  start = NULL,
  test = F
)
}
\arguments{
\item{data}{Dataframe}

\item{col}{Name of the column to pull from}

\item{string}{URL string with "<TEST>" where you want to insert the data}

\item{sample}{number to try}

\item{test}{default to FALSE}
}
\description{
This function creates a table of summary statistics.
}
\examples{
 testInside(y$meetid, col="schoolid", string="http://co.milesplit.com/teams/<TEST>", sample=1, test=F)

}
