% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testingFunctions.R
\name{testOutside}
\alias{testOutside}
\title{Test what's not in the list}
\usage{
testOutside(data, string = "http://co.milesplit.com/meets/<TEST>",
  sample = 20, start = NULL, test = F)
}
\arguments{
\item{data}{Data vector}

\item{string}{URL string with "<TEST>" where you want to insert the data}

\item{sample}{number to try}

\item{test}{default to FALSE}
}
\description{
This function creates a table of summary statistics.
}
\examples{
 testOutside(y$meetid, string="http://co.milesplit.com/meets/<TEST>", sample=20, test=F)

}
