% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cd.r
\name{unique00}
\alias{unique00}
\title{Change Directory}
\usage{
unique00(x, an = T)
}
\arguments{
\item{directory}{The directory you want to set.}
}
\description{
This function allows you to set a directory, whether it exists or not.
}
\examples{
cd()
}
\keyword{directory}
