% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackProjectFunctions.r
\name{wn}
\alias{wn}
\title{Take only numbers before the decimal. "43890.478943" turns to "43890348"}
\usage{
wn(x)
}
\description{
Take only numbers before the decimal. "43890.478943" turns to "43890348"
}
