% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackProjectFunctions.r
\name{wordExists}
\alias{wordExists}
\title{Is this pattern in the string (Logical) - just like grepl}
\usage{
wordExists(pattern, string)
}
\description{
Is this pattern in the string (Logical) - just like grepl
}
