% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackProjectFunctions.r
\name{xvars}
\alias{xvars}
\title{Select variables of a dataset if they exist. Basically, this is to avoid errors for "variable doesn't exist in the dataset".}
\usage{
xvars(data, variables)
}
\description{
Select variables of a dataset if they exist. Basically, this is to avoid errors for "variable doesn't exist in the dataset".
}
