% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cd.r
\name{c2f}
\alias{c2f}
\title{Change Directory}
\usage{
c2f(in_fn, out_fn = in_fn, clean = F)
}
\arguments{
\item{directory}{The directory you want to set.}
}
\description{
This function allows you to set a directory, whether it exists or not.
}
\examples{
cd()
}
\keyword{directory}
