% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackProjectFunctions.r
\name{c2f}
\alias{c2f}
\title{CSV to Feather (if necessary)}
\usage{
c2f(in_fn, out_fn = in_fn, clean = F, pipe = F, txt = F,
  test = F, test_n = 1000, read_in_feather = T)
}
\description{
CSV to Feather (if necessary)
}
