% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackProjectFunctions.r
\name{letters_only}
\alias{letters_only}
\title{Does string contain letters only? If so, it returns TRUE. Otherwise, FALSE.}
\usage{
letters_only(x)
}
\description{
Does string contain letters only? If so, it returns TRUE. Otherwise, FALSE.
}
