% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackProjectFunctions.r
\name{print0}
\alias{print0}
\title{Custom print with systime -- prints whatever you'd like, but with the systime.}
\usage{
print0(x)
}
\description{
Custom print with systime -- prints whatever you'd like, but with the systime.
}
