% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testingFunctions.R, R/trackProjectFunctions.r
\name{read.pipe}
\alias{read.pipe}
\title{Read Pipe Delimited Files}
\usage{
read.pipe(filename, ...)

read.pipe(filename, ...)
}
\arguments{
\item{filename}{filename for pipe-delimited file}
}
\description{
This function creates a table of summary statistics.
}
\examples{
 read.pipe(filename)

}
\keyword{pipe}
\keyword{read}
