% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackProjectFunctions.r
\name{read_feather0}
\alias{read_feather0}
\title{This is a wrapper for the read_feather function. It returns a data.frame and fixes the first name, which often has an extra character/}
\usage{
read_feather0(data, ...)
}
\description{
This is a wrapper for the read_feather function. It returns a data.frame and fixes the first name, which often has an extra character/
}
