% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cd.r
\name{separate0}
\alias{separate0}
\title{Change Directory}
\usage{
separate0(data, col_to_split, sep)
}
\arguments{
\item{directory}{The directory you want to set.}
}
\description{
This function allows you to set a directory, whether it exists or not.
}
\examples{
cd()
}
\keyword{directory}
